+++ database/upgrade/upgrade3.6.0.sql
@@ -35,20 +35,6 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_microgrids_windturbines` (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_microgrids_windturbines_index_1` ON  `myems_system_db`.`tbl_microgrids_windturbines` (`name`);
 
-CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_microgrids_windturbines` (
-  `id` BIGINT NOT NULL AUTO_INCREMENT,
-  `name` VARCHAR(255) NOT NULL,
-  `uuid` CHAR(36) NOT NULL,
-  `microgrid_id` BIGINT NOT NULL,
-  `capacity` DECIMAL(18, 3) NOT NULL,
-  PRIMARY KEY (`id`));
-CREATE INDEX `tbl_microgrids_windturbines_index_1` ON  `myems_system_db`.`tbl_microgrids_windturbines` (`name`);
-
--- ----------------------------
--- Table structure for `myems_user_db`.tbl_verification_codes
--- ----------------------------
-DROP TABLE IF EXISTS `myems_user_db`.`tbl_verification_codes`;
-
 CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_verification_codes`  (
   `id` int NOT NULL AUTO_INCREMENT,
   `recipient_email` VARCHAR(128) NOT NULL,
@@ -58,7 +44,16 @@ CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_verification_codes`  (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_verirication_codes_index_1` ON `myems_user_db`.`tbl_verification_codes` (`recipient_email`, `created_datetime_utc`);
 
+CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_api_keys`  (
+  `id` int NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(128) NOT NULL,
+  `token` VARCHAR(255) NOT NULL,
+  `created_datetime_utc` DATETIME NOT NULL,
+  `expires_datetime_utc` DATETIME NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_api_keys_index_1` ON `myems_user_db`.`tbl_api_keys` (`created_datetime_utc`, `name`);
+
 -- UPDATE VERSION NUMBER
-UPDATE `myems_system_db`.`tbl_versions` SET version='3.6.0RC', release_date='2023-06-18' WHERE id=1;
+UPDATE `myems_system_db`.`tbl_versions` SET version='3.6.0', release_date='2023-06-22' WHERE id=1;
 
 COMMIT;