+++ myems-admin/app/controllers/settings/space/spaceequipment.controller.js
@@ -60,7 +60,8 @@ app.controller('SpaceEquipmentController', function(
 		};
 
 	$scope.getAllEquipments = function() {
-		EquipmentService.getAllEquipments(function (response) {
+        let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		EquipmentService.getAllEquipments(headers, function (response) {
 			if (angular.isDefined(response.status) && response.status === 200) {
 				$scope.equipments = response.data;
 			} else {