+++ myems-api/core/tenanttype.py
@@ -5,7 +5,7 @@ import mysql.connector
 import simplejson as json
 
 import config
-from core.useractivity import user_logger, admin_control
+from core.useractivity import user_logger, admin_control, access_control
 
 
 class TenantTypeCollection:
@@ -20,6 +20,7 @@ class TenantTypeCollection:
 
     @staticmethod
     def on_get(req, resp):
+        access_control(req)
         cnx = mysql.connector.connect(**config.myems_system_db)
         cursor = cnx.cursor()
 
@@ -127,6 +128,7 @@ class TenantTypeItem:
 
     @staticmethod
     def on_get(req, resp, id_):
+        access_control(req)
         if not id_.isdigit() or int(id_) <= 0:
             raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_TENANT_TYPE_ID')